Add disk image creation for Hardware DSL.
With this change, when specifying the storage device to attach to a
storage controller, you can define an image to be created.

e.g.  {:device-type :hard-disk
       :location disk-location
       :create? true
       :size disk-in-mb
       :attachment-type :normal}

The presence of :created? (set to true), :location and :size will prompt
the creation of the disk image. Other optional keys are :variants and
:format (see image/create-medium).