Make manager/power-down more stable.
Before, power-down would return even when the VM was not in a lockable
state. Now it polls for the state until it is lockable. This was
causing calls to destroy after a power down to fail in some instances
because of this race condition. Might be a bug in VBox.